# [router] global tool parser registry