--- a/sgl-router/Cargo.toml
+++ b/sgl-router/Cargo.toml
@@ -64,6 +64,7 @@ prost-types = "0.13"
 deadpool = { version = "0.12", features = ["managed", "rt_tokio_1"] }
 backoff = { version = "0.4", features = ["tokio"] }
 strum = { version = "0.26", features = ["derive"] }
+once_cell = "1.21.3"
 
 [build-dependencies]
 tonic-build = "0.12"